--- ssas/credentials.go
@@ -1,9 +1,0 @@
-package ssas
-
-type Credentials struct {
-	UserID       string
-	ClientID     string
-	ClientSecret string
-	TokenString	 string
-	ClientName   string
-}
